split numba functions
